requirements.txt
pytest==5.4.0
allure-pytest==2.8.15
allure-python-commons==2.8.15
pytest-rerunfailures==9.0
allure-python-commons==2.8.15
configparser==5.0.0
PyYAML==5.3.1
requests==2.23.0
requests-toolbelt==0.9.1
simplejson==3.17.0
ruamel.yaml==0.16.10